Fix permissions on /var/lib/glance/cache
During deployment, it is possible that the folder was previously
created by something else (i.e. in this case, the nspawn deployment
tooling).

We already set the permissions for the subfolders but we don't set
it to the root folder which is causing Glance to fail to start
properly.

Change-Id: I91b3529fd1896ecde814801c0297465f1b956871
(cherry picked from commit 55f9e3600c0c4c87bddb4169b8e5a823cb97dd06)