Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: I63b56c8003efab5a31b9322a865ccbad74022c09
(cherry picked from commit 88334b5a955ced36023b065704a5915fa58f0256)